refactor: Move to package com.appland.appmap.maven
